requirements.txt
Django==3.2
djangorestframework==3.12
mysqlclient==2.0.1
uWSGI==2.0.19.1
djangorestframework-simplejwt==4.7.2
django-cors-headers==3.8.0
django-crispy-forms==1.12.0